Move concurrency handling to a separate class
The new ConcurrentEvaluator is only used for Ontology parsing. It has
all the functionality that the BaseEvaluator had before. The new
BaseEvaluator, however, is the base class for the ProveEvaluator and
does not contain any concurrency handling which renders the concurrency
reladed methods in the ProveEvaluator needless.
