+++ lib/hets/base_evaluator.rb
@@ -1,22 +1,16 @@
 module Hets
 
   # - Meta component
-  # This is the base class of the Node Evaluator It should not be used
+  # This is the base class of the Hets Evaluators It should not be used
   # directly but instead be subclassed. This is necessary since we usually
-  # want to register specific callbacks.  See Hets::NodeEvaluator for the
-  # actual evaluator.
+  # want to register specific callbacks. See Hets::*::*Evaluator for the
+  # actual evaluators.
   class BaseEvaluator
     attr_accessor :hets_evaluator
     attr_accessor :ontology
     attr_accessor :logic_callback
 
-    concurrency_delegates = [
-      :concurrency, :dgnode_stack,
-      :dgnode_stack_id, :next_dgnode_stack_id,
-    ]
-    delegate *concurrency_delegates, to: :hets_evaluator
     delegate :user, to: :hets_evaluator
-    delegate :ontologies_count, to: :hets_evaluator
 
     # - Meta method
     # registers a callback-method for a specific node (with node meaning a
@@ -83,28 +77,5 @@ module Hets
       registrations = self.class.instance_variable_get(:@registrations)
       selected_registrations = registrations[[node_type, order]] || []
     end
-
-    # As concurrency handling is usually performed across
-    # multiple method calls during the parsing-chain,
-    # we will need to initialize and finish concurrency
-    # handling manually. A block-approach is just not
-    # feasible.
-    def initiate_concurrency_handling(ontohub_iri)
-      concurrency.mark_as_processing_or_complain(ontohub_iri,
-        unlock_this_iri: dgnode_stack[dgnode_stack_id])
-      dgnode_stack << ontohub_iri
-    end
-
-    def finish_concurrency_handling
-      all_dgnodes_parsed = next_dgnode_stack_id == hets_evaluator.dgnode_count
-      concurrency.mark_as_finished_processing(dgnode_stack.last) if all_dgnodes_parsed
-    end
-
-    def cancel_concurrency_handling_on_error
-      dgnode_stack.reverse_each do |dgnode|
-        concurrency.unmark_as_processing_on_error(dgnode)
-      end
-    end
-
   end
 end
