+++ lib/hets/prove/evaluator.rb
@@ -38,15 +38,6 @@ module Hets
         self.parser = Parser.new(io || path)
         self.now = Time.now
       end
-
-      # The caller needs those methods to exists, but we don't need the
-      # functionality here. This will be restructured in a later branch.
-      %i(concurrency
-        dgnode_stack
-        dgnode_stack_id
-        next_dgnode_stack_id).each do |method_name|
-        define_method(method_name) {}
-      end
     end
   end
 end
